refactor: rebuild the structure of overall packaging of the project
