--- src/main/java/com/programmers/vouchermanagement/VoucherRepository.java
@@ -1,5 +1,0 @@
-package com.programmers.vouchermanagement;
-
-public interface VoucherRepository {
-    void save(Voucher voucher);
-}
