+++ src/main/java/com/programmers/vouchermanagement/runner/ConsoleAppRunner.java
@@ -1,8 +1,10 @@
-package com.programmers.vouchermanagement;
+package com.programmers.vouchermanagement.runner;
 
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.stereotype.Component;
+
+import com.programmers.vouchermanagement.console.Client;
 
 @Component
 public class ConsoleAppRunner implements ApplicationRunner {
