+++ src/main/java/com/programmers/vouchermanagement/voucher/FixedAmountVoucher.java
@@ -1,6 +1,8 @@
-package com.programmers.vouchermanagement;
+package com.programmers.vouchermanagement.voucher;
 
 import java.util.UUID;
+
+import com.programmers.vouchermanagement.voucher.Voucher;
 
 public class FixedAmountVoucher implements Voucher {
     private final UUID voucherID;
