+++ src/main/java/com/programmers/vouchermanagement/voucher/PercentVoucher.java
@@ -1,6 +1,8 @@
-package com.programmers.vouchermanagement;
+package com.programmers.vouchermanagement.voucher;
 
 import java.util.UUID;
+
+import com.programmers.vouchermanagement.voucher.Voucher;
 
 public class PercentVoucher implements Voucher {
     private final UUID voucherID;
