+++ src/main/java/com/programmers/vouchermanagement/voucher/VoucherRepository.java
@@ -1,0 +1,7 @@
+package com.programmers.vouchermanagement.voucher;
+
+import com.programmers.vouchermanagement.voucher.Voucher;
+
+public interface VoucherRepository {
+    void save(Voucher voucher);
+}
